requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
boto3==1.20.26
botocore==1.23.26
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-storages==1.11.1
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
